pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["shinezeee <h46998600@gmail.com>"]
readme = "README.md"
packages = [{include = "djangoproject"}]

[tool.poetry.dependencies]
python = "^3.13.2"  # Python 3.13은 현재 지원되지 않음
django = "^5.1.6"
psycopg = "^3.2.5"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.5.0"

[tool.poetry.group.dev.dependencies]
black = "<25.1.0"
isort = "<6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"